Minor fixes: * Removed snail food * added formatting comments for emacs * got rid of explicit unboxing

+++ src/lt/pov/FuelLog/CSVImport.java
@@ -1,3 +1,5 @@
+// -*-  c-basic-offset: 4; tab-width: 4; indent-tabs-mode: nil -*-
+
 package lt.pov.FuelLog;
 
 import java.io.BufferedReader;
@@ -10,36 +12,36 @@ import java.util.List;
 
 /**
  * A reader for CSV files of the following format:
- * 
+ *
  * <pre>
  * date,   odometer volume sum    full
  * 2009-02-17,184644,39.00,125.35,1
  * </pre>
- * 
+ *
  * @author alga
  *
  */
 public class CSVImport {
-	private final BufferedReader csv; 
-	
+	private final BufferedReader csv;
+
 	public CSVImport(String filename) throws FileNotFoundException {
 		csv = new BufferedReader(new FileReader(filename));
 	}
-	
+
 	public List<Record> entries() throws IOException {
 		List<Record> result = new ArrayList<Record>();
 		String line;
 		while ((line = csv.readLine()) != null) {
 			String[] segments = line.trim().split(",");
-			result.add(new Record(segments[0], 
-					              Integer.valueOf(segments[1]).intValue(),
-					              Double.valueOf(segments[2]).doubleValue(),
-					              Double.valueOf(segments[3]).doubleValue(),
+			result.add(new Record(segments[0],
+                                  Integer.valueOf(segments[1]),
+					              Double.valueOf(segments[2]),
+					              Double.valueOf(segments[3]),
 					              segments[4].equals("1")));
 		}
 		return result;
 	}
-	
+
 	public static class Record {
 		final String date;
 		final int odometer;
